Make example fit in 1 file by using plugin package
