+++ cached_network_image/example/lib/main.dart
@@ -1,151 +1,204 @@
+import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
-
-import 'template/globals.dart';
+import 'package:baseflow_plugin_template/baseflow_plugin_template.dart';
+import 'package:flutter_blurhash/flutter_blurhash.dart';
 
 void main() {
-  runApp(BaseflowPluginExample());
+  runApp(BaseflowPluginExample(
+    pluginName: 'CachedNetworkImage',
+    githubURL: 'https://github.com/Baseflow/flutter_cache_manager',
+    pubDevURL: 'https://pub.dev/packages/flutter_cache_manager',
+    pages: [
+      BasicContent.createPage(),
+      ListContent.createPage(),
+      GridContent.createPage(),
+    ],
+  ));
 }
 
-/// A Flutter application demonstrating the functionality of this plugin
-class BaseflowPluginExample extends StatelessWidget {
-  /// [MaterialColor] to be used in the app [ThemeData]
-  final MaterialColor themeMaterialColor =
-      createMaterialColor(const Color.fromRGBO(48, 49, 60, 1));
+/// Demonstrates a [StatelessWidget] containing [CachedNetworkImage]
+class BasicContent extends StatelessWidget {
+  static ExamplePage createPage() {
+    return ExamplePage(Icons.image, (context) => BasicContent());
+  }
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Baseflow $pluginName',
-      theme: ThemeData(
-        accentColor: Colors.white60,
-        backgroundColor: const Color.fromRGBO(48, 49, 60, 0.8),
-        buttonTheme: ButtonThemeData(
-          buttonColor: themeMaterialColor.shade500,
-          disabledColor: themeMaterialColor.withRed(200),
-          splashColor: themeMaterialColor.shade50,
-          textTheme: ButtonTextTheme.primary,
-        ),
-        bottomAppBarColor: const Color.fromRGBO(57, 58, 71, 1),
-        hintColor: themeMaterialColor.shade500,
-        primarySwatch: createMaterialColor(const Color.fromRGBO(48, 49, 60, 1)),
-        textTheme: TextTheme(
-          bodyText1: TextStyle(
-            color: Colors.white,
-            fontSize: 16,
-            height: 1.3,
-          ),
-          bodyText2: TextStyle(
-            color: Colors.white,
-            fontSize: 18,
-            height: 1.2,
-          ),
-          button: TextStyle(color: Colors.white),
-          headline1: TextStyle(
-            color: Colors.white,
-            fontSize: 18,
-          ),
-        ),
-        visualDensity: VisualDensity.adaptivePlatformDensity,
-        inputDecorationTheme: const InputDecorationTheme(
-          fillColor: Color.fromRGBO(37, 37, 37, 1),
-          filled: true,
-        ),
-      ),
-      home: AppHome(),
-    );
-  }
-
-  /// Creates a [MaterialColor] based on the supplied [Color]
-  static MaterialColor createMaterialColor(Color color) {
-    List strengths = <double>[.05];
-    final swatch = <int, Color>{};
-    final r = color.red, g = color.green, b = color.blue;
-
-    for (var i = 1; i < 10; i++) {
-      strengths.add(0.1 * i);
-    }
-    for (var strength in strengths) {
-      final ds = 0.5 - strength;
-      swatch[(strength * 1000).round()] = Color.fromRGBO(
-        r + ((ds < 0 ? r : (255 - r)) * ds).round(),
-        g + ((ds < 0 ? g : (255 - g)) * ds).round(),
-        b + ((ds < 0 ? b : (255 - b)) * ds).round(),
-        1,
-      );
-    }
-    return MaterialColor(color.value, swatch);
-  }
-}
-
-/// A Flutter example demonstrating how the [pluginName] plugin could be used
-class AppHome extends StatefulWidget {
-  /// Constructs the [AppHome] class
-  AppHome({Key? key}) : super(key: key);
-
-  @override
-  _AppHomeState createState() => _AppHomeState();
-}
-
-class _AppHomeState extends State<AppHome> {
-  static final PageController _pageController = PageController(initialPage: 0);
-  int _currentPage = 0;
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Theme.of(context).bottomAppBarColor,
-        title: Center(
-          child: Image.asset(
-            'res/images/baseflow_logo_def_light-02.png',
-            width: 140,
-          ),
-        ),
-      ),
-      backgroundColor: Theme.of(context).backgroundColor,
-      body: PageView(
-        controller: _pageController,
-        children: pages,
-        onPageChanged: (page) {
-          setState(() {
-            _currentPage = page;
-          });
-        },
-      ),
-      bottomNavigationBar: _bottomAppBar(),
-    );
-  }
-
-  BottomAppBar _bottomAppBar() {
-    return BottomAppBar(
-      elevation: 5,
-      color: Theme.of(context).bottomAppBarColor,
-      child: Row(
-        mainAxisSize: MainAxisSize.max,
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: List.unmodifiable(() sync* {
-          for (var i = 0; i < pages.length; i++) {
-            yield Expanded(
-              child: IconButton(
-                iconSize: 30,
-                icon: Icon(icons.elementAt(i)),
-                color: _bottomAppBarIconColor(i),
-                onPressed: () => _animateToPage(i),
+    return SingleChildScrollView(
+      child: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            _blurHashImage(),
+            _sizedContainer(
+              const Image(
+                image: CachedNetworkImageProvider(
+                  'https://via.placeholder.com/350x150',
+                ),
               ),
-            );
-          }
-        }()),
+            ),
+            _sizedContainer(
+              CachedNetworkImage(
+                progressIndicatorBuilder: (context, url, progress) => Center(
+                  child: CircularProgressIndicator(
+                    value: progress.progress,
+                  ),
+                ),
+                imageUrl:
+                    'https://images.unsplash.com/photo-1532264523420-881a47db012d?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9',
+              ),
+            ),
+            _sizedContainer(
+              CachedNetworkImage(
+                placeholder: (context, url) =>
+                    const CircularProgressIndicator(),
+                imageUrl: 'https://via.placeholder.com/200x150',
+              ),
+            ),
+            _sizedContainer(
+              CachedNetworkImage(
+                imageUrl: 'https://via.placeholder.com/300x150',
+                imageBuilder: (context, imageProvider) => Container(
+                  decoration: BoxDecoration(
+                    image: DecorationImage(
+                      image: imageProvider,
+                      fit: BoxFit.cover,
+                      colorFilter: const ColorFilter.mode(
+                        Colors.red,
+                        BlendMode.colorBurn,
+                      ),
+                    ),
+                  ),
+                ),
+                placeholder: (context, url) =>
+                    const CircularProgressIndicator(),
+                errorWidget: (context, url, error) => const Icon(Icons.error),
+              ),
+            ),
+            CachedNetworkImage(
+              imageUrl: 'https://via.placeholder.com/300x300',
+              placeholder: (context, url) => const CircleAvatar(
+                backgroundColor: Colors.amber,
+                radius: 150,
+              ),
+              imageBuilder: (context, image) => CircleAvatar(
+                backgroundImage: image,
+                radius: 150,
+              ),
+            ),
+            _sizedContainer(
+              CachedNetworkImage(
+                imageUrl: 'https://notAvalid.uri',
+                placeholder: (context, url) =>
+                    const CircularProgressIndicator(),
+                errorWidget: (context, url, error) => const Icon(Icons.error),
+              ),
+            ),
+            _sizedContainer(
+              CachedNetworkImage(
+                imageUrl: 'not a uri at all',
+                placeholder: (context, url) =>
+                    const CircularProgressIndicator(),
+                errorWidget: (context, url, error) => const Icon(Icons.error),
+              ),
+            ),
+            _sizedContainer(
+              CachedNetworkImage(
+                maxHeightDiskCache: 10,
+                imageUrl: 'https://via.placeholder.com/350x200',
+                placeholder: (context, url) =>
+                    const CircularProgressIndicator(),
+                errorWidget: (context, url, error) => const Icon(Icons.error),
+                fadeOutDuration: const Duration(seconds: 1),
+                fadeInDuration: const Duration(seconds: 3),
+              ),
+            ),
+          ],
+        ),
       ),
     );
   }
 
-  void _animateToPage(int page) {
-    _pageController.animateToPage(page,
-        duration: const Duration(milliseconds: 200), curve: Curves.linear);
+  Widget _blurHashImage() {
+    return SizedBox(
+      width: double.infinity,
+      child: CachedNetworkImage(
+        placeholder: (context, url) => const AspectRatio(
+          aspectRatio: 1.6,
+          child: BlurHash(hash: 'LEHV6nWB2yk8pyo0adR*.7kCMdnj'),
+        ),
+        imageUrl: 'https://blurha.sh/assets/images/img1.jpg',
+        fit: BoxFit.cover,
+      ),
+    );
   }
 
-  Color _bottomAppBarIconColor(int page) {
-    return _currentPage == page ? Colors.white : Theme.of(context).accentColor;
+  Widget _sizedContainer(Widget child) {
+    return SizedBox(
+      width: 300.0,
+      height: 150.0,
+      child: Center(child: child),
+    );
+  }
+}
+
+/// Demonstrates a [ListView] containing [CachedNetworkImage]
+class ListContent extends StatelessWidget {
+  static ExamplePage createPage() {
+    return ExamplePage(Icons.list, (context) => ListContent());
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return ListView.builder(
+      itemBuilder: (BuildContext context, int index) => Card(
+        child: Column(
+          children: <Widget>[
+            CachedNetworkImage(
+              imageUrl: 'https://loremflickr.com/320/240/music?lock=$index',
+              placeholder: (BuildContext context, String url) => Container(
+                width: 320,
+                height: 240,
+                color: Colors.purple,
+              ),
+            ),
+          ],
+        ),
+      ),
+      itemCount: 250,
+    );
+  }
+}
+
+/// Demonstrates a [GridView] containing [CachedNetworkImage]
+class GridContent extends StatelessWidget {
+  static ExamplePage createPage() {
+    return ExamplePage(Icons.grid_on, (context) => GridContent());
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return GridView.builder(
+      itemCount: 250,
+      gridDelegate:
+          const SliverGridDelegateWithFixedCrossAxisCount(crossAxisCount: 3),
+      itemBuilder: (BuildContext context, int index) => CachedNetworkImage(
+        imageUrl: 'https://loremflickr.com/100/100/music?lock=$index',
+        placeholder: _loader,
+        errorWidget: _error,
+      ),
+    );
+  }
+
+  Widget _loader(BuildContext context, String url) {
+    return const Center(
+      child: CircularProgressIndicator(),
+    );
+  }
+
+  Widget _error(BuildContext context, String url, dynamic error) {
+    print(error);
+    return const Center(child: Icon(Icons.error));
   }
 }
